In Progress

+++ TestSutraUsers/src/main/java/com/tathvatech/survey/response/SurveyResponse.java
@@ -10,8 +10,15 @@ import java.util.Date;
 import java.util.HashMap;
 import java.util.Iterator;
 
+import com.tathvatech.forms.oid.FormResponseOID;
+import com.tathvatech.forms.response.FormResponseStats;
 import com.tathvatech.forms.response.ResponseMasterNew;
 import com.tathvatech.survey.common.SurveyDefinition;
+import com.tathvatech.survey.entity.ResponseFlags;
+import com.tathvatech.survey.inf.SurveyItemBase;
+import com.tathvatech.survey.inf.SurveySaveItemBase;
+import com.tathvatech.unit.common.UnitQuery;
+import com.tathvatech.unit.entity.UnitTestProc;
 import com.tathvatech.user.entity.User;
 
 
@@ -73,7 +80,7 @@ public class SurveyResponse
 		this(sd);
 		responseStartTime = new Date();
 		ipaddress = "";
-		testProcPk = unitForm.getPk();
+		testProcPk = (int) unitForm.getPk();
 		userPk = (int) user.getPk();
 		this.user = user;
 		status = ResponseMasterNew.STATUS_INPROGRESS;
